Move RecoverDefault into profiles and apply it on all discovery layers
