+++ linkerd/service-profiles/src/lib.rs
@@ -10,6 +10,7 @@ use std::{
 use tower::util::{Oneshot, ServiceExt};
 
 mod client;
+pub mod default;
 pub mod discover;
 pub mod http;
 pub mod split;
